requirements.txt
Flask
flask_login
flask_sqlalchemy
flask_wtf
Pillow
Requests
SQLAlchemy
ultralytics
WTForms
pymysql